# contracts/admin-api.openapi.yaml
openapi: 3.1.0
info:
  title: AI Control Layer – Admin API
  description: Management and reporting API consumed by the admin panel.
  version: 1.0.0
  x-gateway-version: 0.1.0
paths:
...
  /admin/v1/users:
    get:
      tags:
      - access
      summary: List Users
      operationId: listUsers
      security:
      - HTTPBearer: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: group
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Group
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
                title: Response Listusers
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/users/{user_id}:
    get:
      tags:
      - access
      summary: Get User
      operationId: getUser
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/users/{user_id}/effective-access:
    get:
      tags:
      - access
      summary: Effective Access
      description: 'Resolved access with sources: org locks → group grants → user grants (with expiry).'
      operationId: getEffectiveAccess
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EffectiveAccess'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/users/{user_id}/activity:
    get:
      tags:
      - access
      summary: Activity
      operationId: getUserActivity
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Since
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventSummary'
                title: Response Getuseractivity
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/users/{user_id}/breakglass:
    post:
      tags:
      - access
      summary: Break Glass
      description: View raw content of one event. Requires a reason; audited as its own `breakglass` event.
      operationId: breakGlass
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BreakGlassRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BreakGlassResponse'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/groups:
    get:
      tags:
      - access
      summary: List Groups
      operationId: listGroups
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Group'
                type: array
                title: Response Listgroups
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
  /admin/v1/grants:
    get:
      tags:
      - access
      summary: List Grants
      operationId: listGrants
      security:
      - HTTPBearer: []
      parameters:
      - name: subject
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
      - name: resource_type
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/GrantResourceType'
          - type: 'null'
          title: Resource Type
      - name: resource
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource
      - name: active
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          default: true
          title: Active
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Grant'
                title: Response Listgrants
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - access
      summary: Create Grant
      description: Create a per-user/group grant. Rejected (422) if an org lock forbids it outright.
      operationId: createGrant
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/grants/{grant_id}:
    delete:
      tags:
      - access
      summary: Revoke Grant
      operationId: revokeGrant
      security:
      - HTTPBearer: []
      parameters:
      - name: grant_id
        in: path
        required: true
        schema:
          type: string
          title: Grant Id
      - name: reason
        in: query
        required: true
        schema:
          type: string
          title: Reason
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/grants/changes:
    get:
      tags:
      - access
      summary: Grant Changes
      operationId: listGrantChanges
      security:
      - HTTPBearer: []
      parameters:
      - name: subject
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrantChange'
                title: Response Listgrantchanges
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/users/{user_id}/api-keys:
    get:
      tags:
      - access
      summary: List Keys
      operationId: listApiKeys
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
                title: Response Listapikeys
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - access
      summary: Create Key
      operationId: createApiKey
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreated'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/api-keys/{key_id}:
    delete:
      tags:
      - access
      summary: Revoke Key
      operationId: revokeApiKey
      security:
      - HTTPBearer: []
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
          title: Key Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/v1/me:
    get:
      tags:
      - access
      summary: Me
      description: "Employee self-service: the caller's own effective access."
      operationId: getMyAccess
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EffectiveAccess'
        '401':
          description: Missing or invalid credentials
        '403':
          description: Insufficient role / forbidden resource
        '501':
          description: Not implemented yet
      security:
      - HTTPBearer: []
...
components:
  schemas:
    Action:
      type: string
      enum:
      - allow
      - monitor
      - redact
      - pseudonymise
      - sanitize
      - route_local
      - downgrade
      - require_approval
      - block
      title: Action
      description: Decision actions (concept §6.3).
    ApiKey:
      properties:
        id:
          type: string
          title: Id
        user_id:
          type: string
          title: User Id
        name:
          type: string
          title: Name
        prefix:
          type: string
          title: Prefix
          description: First characters of the key, for identification.
        created_at:
          type: string
          format: date-time
          title: Created At
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        last_used_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Used At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
      additionalProperties: false
      type: object
      required:
      - id
      - user_id
      - name
      - prefix
      - created_at
      title: ApiKey
    ApiKeyCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
      additionalProperties: false
      type: object
      required:
      - name
      title: ApiKeyCreate
    ApiKeyCreated:
      properties:
        id:
          type: string
          title: Id
        user_id:
          type: string
          title: User Id
        name:
          type: string
          title: Name
        prefix:
          type: string
          title: Prefix
          description: First characters of the key, for identification.
        created_at:
          type: string
          format: date-time
          title: Created At
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        last_used_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Used At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
        key:
          type: string
          title: Key
          description: Full key. Returned exactly once; only a hash is stored.
      additionalProperties: false
      type: object
      required:
      - id
      - user_id
      - name
      - prefix
      - created_at
      - key
      title: ApiKeyCreated
...
    BreakGlassRequest:
      properties:
        event_id:
          type: string
          title: Event Id
        reason:
          type: string
          maxLength: 1000
          minLength: 10
          title: Reason
      additionalProperties: false
      type: object
      required:
      - event_id
      - reason
      title: BreakGlassRequest
    BreakGlassResponse:
      properties:
        event_id:
          type: string
          title: Event Id
        audit_event_id:
          type: string
          title: Audit Event Id
        raw_payload:
          anyOf:
          - type: string
          - type: 'null'
          title: Raw Payload
          description: Null when raw retention is disabled or expired.
        available:
          type: boolean
          title: Available
      additionalProperties: false
      type: object
      required:
      - event_id
      - audit_event_id
      - raw_payload
      - available
      title: BreakGlassResponse
...
    DataClass:
      type: string
      enum:
      - public
      - internal
      - confidential
      - restricted
      title: DataClass
      description: Confidentiality classes, ordered from least to most sensitive.
...
    EffectiveAccess:
      properties:
        subject:
          type: string
          title: Subject
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        groups:
          items:
            type: string
          type: array
          title: Groups
        preset:
          $ref: '#/components/schemas/Preset'
        preset_source:
          type: string
          title: Preset Source
        items:
          items:
            $ref: '#/components/schemas/EffectiveAccessItem'
          type: array
          title: Items
        budgets:
          additionalProperties:
            additionalProperties:
              type: number
            type: object
          type: object
          title: Budgets
        policy_version:
          type: string
          title: Policy Version
        grants_version:
          type: string
          title: Grants Version
      additionalProperties: false
      type: object
      required:
      - subject
      - groups
      - preset
      - preset_source
      - items
      - policy_version
      - grants_version
      title: EffectiveAccess
    EffectiveAccessItem:
      properties:
        resource_type:
          $ref: '#/components/schemas/GrantResourceType'
        resource:
          type: string
          title: Resource
        effect:
          type: string
          enum:
          - allow
          - deny
          title: Effect
        source:
          type: string
          enum:
          - org_lock
          - group
          - user
          - default
          title: Source
        source_ref:
          type: string
          title: Source Ref
          description: Lock id, group name, or grant id.
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        constraints:
          $ref: '#/components/schemas/GrantConstraints'
        capped_by_lock:
          anyOf:
          - type: string
          - type: 'null'
          title: Capped By Lock
          description: Org lock that narrowed this grant, if any.
      additionalProperties: false
      type: object
      required:
      - resource_type
      - resource
      - effect
      - source
      - source_ref
      title: EffectiveAccessItem
...
    EventSummary:
      properties:
        event_id:
          type: string
          title: Event Id
        seq:
          type: integer
          title: Seq
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        event_type:
          type: string
          title: Event Type
        severity:
          $ref: '#/components/schemas/Severity'
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
        session_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Id
        subject:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
        groups:
          items:
            type: string
          type: array
          title: Groups
        agent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Id
        point:
          anyOf:
          - $ref: '#/components/schemas/InspectionPoint'
          - type: 'null'
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        tool:
          anyOf:
          - type: string
          - type: 'null'
          title: Tool
        action:
          anyOf:
          - $ref: '#/components/schemas/Action'
          - type: 'null'
        rule_ids:
          items:
            type: string
            pattern: ^[A-Z][A-Z0-9]*(-[A-Z0-9_.]+)+$
            description: Rule identifier (control id, or control id + sub-rule, or signature id).
          type: array
          title: Rule Ids
        risk_score:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Risk Score
        latency_ms:
          anyOf:
          - type: number
          - type: 'null'
          title: Latency Ms
      additionalProperties: false
      type: object
      required:
      - event_id
      - seq
      - timestamp
      - event_type
      - severity
      title: EventSummary
...
    Grant:
      properties:
        subject_type:
          type: string
          enum:
          - user
          - group
          title: Subject Type
        subject:
          type: string
          title: Subject
          description: User subject/username or group path.
        resource_type:
          $ref: '#/components/schemas/GrantResourceType'
        resource:
          type: string
          title: Resource
          examples:
          - smart
          - gemini/flash
          - opencode.bash
          - mcp:jira
        effect:
          type: string
          enum:
          - allow
          - deny
          title: Effect
          default: allow
        constraints:
          $ref: '#/components/schemas/GrantConstraints'
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        reason:
          type: string
          maxLength: 500
          minLength: 3
          title: Reason
        id:
          type: string
          title: Id
        created_by:
          type: string
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
        revoked_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Revoked By
        active:
          type: boolean
          title: Active
      additionalProperties: false
      type: object
      required:
      - subject_type
      - subject
      - resource_type
      - resource
      - reason
      - id
      - created_by
      - created_at
      - active
      title: Grant
    GrantChange:
      properties:
        id:
          type: integer
          title: Id
        grant_id:
          type: string
          title: Grant Id
        change:
          type: string
          enum:
          - create
          - revoke
          - expire
          title: Change
        actor:
          type: string
          title: Actor
        reason:
          type: string
          title: Reason
        at:
          type: string
          format: date-time
          title: At
        snapshot:
          $ref: '#/components/schemas/Grant'
      additionalProperties: false
      type: object
      required:
      - id
      - grant_id
      - change
      - actor
      - reason
      - at
      - snapshot
      title: GrantChange
    GrantConstraints:
      properties:
        data_classes:
          anyOf:
          - items:
              $ref: '#/components/schemas/DataClass'
            type: array
          - type: 'null'
          title: Data Classes
        budget_share:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Budget Share
        preset:
          anyOf:
          - $ref: '#/components/schemas/Preset'
          - type: 'null'
      additionalProperties: false
      type: object
      title: GrantConstraints
    GrantCreate:
      properties:
        subject_type:
          type: string
          enum:
          - user
          - group
          title: Subject Type
        subject:
          type: string
          title: Subject
          description: User subject/username or group path.
        resource_type:
          $ref: '#/components/schemas/GrantResourceType'
        resource:
          type: string
          title: Resource
          examples:
          - smart
          - gemini/flash
          - opencode.bash
          - mcp:jira
        effect:
          type: string
          enum:
          - allow
          - deny
          title: Effect
          default: allow
        constraints:
          $ref: '#/components/schemas/GrantConstraints'
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        reason:
          type: string
          maxLength: 500
          minLength: 3
          title: Reason
      additionalProperties: false
      type: object
      required:
      - subject_type
      - subject
      - resource_type
      - resource
      - reason
      title: GrantCreate
    GrantResourceType:
      type: string
      enum:
      - model
      - alias
      - connector
      - skill
      - tool
      - mcp_server
      title: GrantResourceType
    Group:
      properties:
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
          default: ''
        preset:
          anyOf:
          - $ref: '#/components/schemas/Preset'
          - type: 'null'
        members:
          type: integer
          title: Members
          default: 0
        source:
          type: string
          enum:
          - keycloak
          - policy
          - both
          title: Source
          default: both
      additionalProperties: false
      type: object
      required:
      - name
      title: Group
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InspectionPoint:
      type: string
      enum:
      - ingress
      - egress
      - tool_call
      - tool_result
      - embeddings
      - agent_message
      - artifact_load
      - mcp_initialize
      - mcp_tools_list
      title: InspectionPoint
      description: Where in the traffic flow a payload is inspected (concept §6.1).
...
    Preset:
      type: string
      enum:
      - monitor
      - balanced
      - strict
      - paranoid
      title: Preset
    PrincipalKind:
      type: string
      enum:
      - user
      - agent
      - service
      - anonymous
      title: PrincipalKind
...
    Severity:
      type: string
      enum:
      - info
      - low
      - medium
      - high
      - critical
      title: Severity
...
    User:
      properties:
        id:
          type: string
          title: Id
        subject:
          type: string
          title: Subject
        username:
          type: string
          title: Username
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        kind:
          $ref: '#/components/schemas/PrincipalKind'
          default: user
        groups:
          items:
            type: string
          type: array
          title: Groups
        roles:
          items:
            type: string
          type: array
          title: Roles
        first_seen:
          type: string
          format: date-time
          title: First Seen
        last_seen:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Seen
        disabled:
          type: boolean
          title: Disabled
          default: false
      additionalProperties: false
      type: object
      required:
      - id
      - subject
      - username
      - first_seen
      title: User
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      description: Keycloak access token (JWT) or personal API key (`acl_...`).
      scheme: bearer
servers:
- url: http://localhost:8000